refactor: Create verbs/ in annotations
